+++ src/main/java/org/ject/support/admin/member/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package org.ject.support.domain.admin.dto;
+package org.ject.support.admin.member.dto;
 
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Builder;
