+++ src/main/java/org/ject/support/domain/member/repository/MemberQueryRepository.java
@@ -2,7 +2,7 @@ package org.ject.support.domain.member.repository;
 
 import org.ject.support.domain.member.JobFamily;
 import org.ject.support.domain.member.Role;
-import org.ject.support.domain.admin.dto.MemberResponse;
+import org.ject.support.admin.member.dto.MemberResponse;
 import org.ject.support.domain.member.dto.TeamMemberNames;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
