+++ src/main/java/org/ject/support/domain/member/repository/MemberQueryRepositoryImpl.java
@@ -6,10 +6,10 @@ import com.querydsl.core.types.dsl.CaseBuilder;
 import com.querydsl.jpa.JPQLQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.ject.support.common.data.PageResponse;
-import org.ject.support.domain.admin.dto.QMemberResponse;
+import org.ject.support.admin.member.dto.QMemberResponse;
 import org.ject.support.domain.member.JobFamily;
 import org.ject.support.domain.member.Role;
-import org.ject.support.domain.admin.dto.MemberResponse;
+import org.ject.support.admin.member.dto.MemberResponse;
 import org.ject.support.domain.member.dto.QTeamMemberNames;
 import org.ject.support.domain.member.dto.TeamMemberNames;
 import org.springframework.data.domain.Page;
