+++ src/test/java/org/ject/support/admin/member/controller/MemberManagementControllerTest.java
@@ -1,4 +1,4 @@
-package org.ject.support.domain.admin.controller;
+package org.ject.support.admin.member.controller;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -16,15 +16,16 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
+import org.ject.support.admin.member.controller.MemberManagementController;
 import org.ject.support.base.UnitTestSupport;
-import org.ject.support.domain.admin.service.MemberManagementService;
+import org.ject.support.admin.member.service.MemberManagementService;
 import org.ject.support.domain.member.JobFamily;
 import org.ject.support.domain.member.Role;
-import org.ject.support.domain.admin.dto.MemberBulkDeleteRequest;
-import org.ject.support.domain.admin.dto.MemberDetailResponse;
-import org.ject.support.domain.admin.dto.MemberRegisterRequest;
-import org.ject.support.domain.admin.dto.MemberResponse;
-import org.ject.support.domain.admin.dto.MemberEditRequest;
+import org.ject.support.admin.member.dto.MemberBulkDeleteRequest;
+import org.ject.support.admin.member.dto.MemberDetailResponse;
+import org.ject.support.admin.member.dto.MemberRegisterRequest;
+import org.ject.support.admin.member.dto.MemberResponse;
+import org.ject.support.admin.member.dto.MemberEditRequest;
 import org.ject.support.domain.member.exception.MemberErrorCode;
 import org.ject.support.domain.member.exception.MemberException;
 import org.junit.jupiter.api.BeforeEach;
