+++ src/test/java/org/ject/support/admin/member/service/MemberManagementServiceTest.java
@@ -1,4 +1,4 @@
-package org.ject.support.domain.admin.service;
+package org.ject.support.admin.member.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -9,10 +9,11 @@ import static org.mockito.Mockito.verify;
 
 import java.util.List;
 import java.util.Optional;
+import org.ject.support.admin.member.service.MemberManagementService;
 import org.ject.support.base.UnitTestSupport;
-import org.ject.support.domain.admin.dto.MemberEditRequest;
-import org.ject.support.domain.admin.dto.MemberRegisterRequest;
-import org.ject.support.domain.admin.dto.MemberResponse;
+import org.ject.support.admin.member.dto.MemberEditRequest;
+import org.ject.support.admin.member.dto.MemberRegisterRequest;
+import org.ject.support.admin.member.dto.MemberResponse;
 import org.ject.support.domain.member.JobFamily;
 import org.ject.support.domain.member.Role;
 import org.ject.support.domain.member.entity.Member;
